zPAM3.30 PREVIEW
News:
[3.4.16] New pam mode names - cg was removed, now only pub and comp modes are available with sub modes.
	- modes: pub, comp, warmup
	- sub-modes: 2v2, rifle, russian, lan, pcw
	- Examples:
		- /rcon pam_mode pub  			- Public mode (custom settings)
		- /rcon pam_mode comp  			- Competitive mode
		- /rcon pam_mode comp_mr3  		- Competitive mode with max round 4
		- /rcon pam_mode comp_pcw   		- Competitive mode for team mix
		- /rcon pam_mode comp_pcw_russian 	- Competitive mode for team mix, russian side on russian maps
		- /rcon pam_mode comp_2v2 		- Competitive mode, scope and shotgun disabled
		- /rcon pam_mode comp_rifle 		- Competitive mode, bolt action and sniper only
		- /rcon pam_mode comp_rifle_2v2		- Competitive mode, bolt action only, sniper disabled
		- /rcon pam_mode comp_lan  		- Competitive mode on LAN, no readyup time limits
[3.4.17] Added support of HQ, CTF, HTF and RE gametypes
[3.4.18] Added rifle only mode
[3.3.44] Cvar system rewrited; any change to server settings is retained even when map changes; new cvar /pam_mode_custom is defined - it tells that changes made to server settings stays between map
[3.3.45] Warnings about wrong server settings (no password, cheats enabled, punkbuster disabled, cvars changes) is changed; its showed in left top corner under the score; if some of the server settings is changed, detailed list of changed cvars is showed; punkbuster warning is removed
[3.3.46] Hand hit box fix - if hands are in front of body and game somehow badly interprets it as hit to the hand, PAM change it to hit to the body; its an extension to already existing fix in 3.22 for left hand - now its applies also for right hand
[3.3.47] Torso hitbox fix; lower torso hitbox (pelvis area) is slightly enlarged as workaround for bad hitbox registration between torso_lower and right/left_leg_upper; this change effectively applies only for rifles, because other weapons has the same damage for torso_lower and right/left_leg_upper
[3.3.48] Consistent shotgun - this is the new name for new shotgun; this shotgun fixes close range hits; it replaces rebalanced shotgun
[3.3.49] Ladder weapon bug fix - silent use of ladder is no more possible; its a situation when you double scroll your weapon, hold fire button and use a ladder
[3.3.50] Sliding bug fix; g_knockback is set to 0; when you are hit, no move energy is applied (knockback); this should fix fast sliding on ground / getting bugged into ground
[3.3.51] Sniper shotgun info - weapons of players with sniper and shotgun are showed next to compas at start and the end of the round
[3.3.52] Colored time - color of round timer changes to orange when 30 seconds are left and red when 15 seconds are left
[3.3.53] Damage feedback (hit cross in center) - slightly changed to better distinguish between hit and kill; if its hit only, cross is more transparent and disappears quicker; if its kill, cross is fully showed and stays longer
[3.3.54] New blood effect to better visualize hit location; can be enabled / disabled via /cg_blood
[3.3.55] RCON menu - new menu for changing map, server settings and player kick; is showed in all menus in new place
[3.3.56] Auto-spectator - better auto-switch algorithm; hit info; player names; auto hide of help text
[3.3.57] Auto-spectator - automatic killcam replay based on interesting actions (multikills, nozoom)
[3.3.58] Match info - team names now can contains special characters like # _ ! < > etc..
[3.3.59] Bash mode text over the screen is showed when bash mode is called.
[3.3.60] Cvar warning all over the screen is removed for cvar developer and polygons
[3.3.61] Objective text showed in scoreboard is removed
[3.3.62] SD voiceover sound "Move in!" removed at the start of the round
[3.3.63] SD voiceover sound when bomb is planted is shortened - only "Explosives planted" sound is retained
[3.3.64] Strat - up to 9 bot path recording can be saved; up to 9 bot path recordings can be played at once; automatic record stop is now faster
[3.3.65] New mappack file zpam_maps_v1.iwd with maps Toujane FIX v2, Burgundy FIX v1, Dawnville FIX v2, Matmata FIX v2, Carentan FIX v2, Chelm, Vallente, wawa_3dAim
[3.3.66] Timeout settings changes:
	- timeouts per map: 4 (previously 2)
	- timeouts per side: 2 (previously 1)
	- timeout length: 2min (previously 5min)
	- so total timeout time would be 8min per map (previously 10min)
	- overtimes does not reset called timeouts count - it means if you called out all of the timeouts in standard game time, you dont get new timeouts in overtime. (same as previous)
	- when timeout time expires, "Time is over" warning message appears - setting one team ready will end the timeout. (previously it automatically resumes the game)
[3.3.67] Time between halfs and maps changes:
	- between maps: 5min (as previous)
	- between halfs: 2min (previously 5min)
	- between halfs at overtime: 1min (previously 5min)
	- for LAN mode, there will be no time limits between halfs and maps

Bug fixes:
[2.3.5] Error while running server on 1.0 version fixed
[2.3.6] Fixed bug when you drop MG and step sound was made when mg peek fix was enabled
[2.3.7] Matchinfo - fixed bug when team names does not match players in teams; happens when pistol bash was called and players swap teams
[2.3.8] Auto-recording - fixed bug with empty team names when bash mode was called
[2.3.9] Auto-recording - server crash when somebody joins a team at the end of the round (only windows servers) -> fixed
[2.3.10] Auto-recording - wont start if main menu is opened, quick messages menu is opened or game is minimized -> fixed
[2.3.11] Auto-spectator - When replaying killcam in spectator, name of spectated player is not removed when the killcam ends -> fixed
[2.3.12] Fixed text PC_PATCH_1_1_SD_OBJECTIVES in SD server info
[2.3.13] Weapon name next to ammo HUD is not visible when enemy player list is enabled -> fixed by showing the weapon name under the ammo
[1.14] BAR sound bug - when BAR weapon is picked from ground, it makes pickup sound for every player -> fixed

+++ source/maps/mp/gametypes/global/player.gsc
@@ -21,6 +21,7 @@ IsMoving()
 	self endon("disconnect");
 
 	self.isMoving = false;
+	self.isWaling = false;
 	self.movingDifference = 0;
 
 	origin = self getOrigin();
@@ -42,8 +43,22 @@ IsMoving()
 		else
 		{/*
 		  if (self.isMoving)
-			self iprintln("^1Stopped");*/
+			self iprintln("^1Stopped moving");*/
 		  self.isMoving = false;
+		}
+
+
+		if (self.movingDifference > 10.0)
+		{/*
+		  if (!self.isWaling)
+			self iprintln("Walking");*/
+		  self.isWaling = true;
+		}
+		else
+		{/*
+		  if (self.isWaling)
+			self iprintln("^1Stopped walking");*/
+		  self.isWaling = false;
 		}
 
 		origin_old = origin;
@@ -61,28 +76,34 @@ AddTags()
 	// Spawn help tag that will be linked to head
 	self.headTag = spawn("script_origin",(0,0,0));
 	//self.headTag thread maps\mp\gametypes\global\developer::showWaypoint();
+	self.pelvisTag = spawn("script_origin",(0,0,0));
+	//self.pelvisTag thread maps\mp\gametypes\global\developer::showWaypoint();
 
-	wait level.frame;
-
+	dead = true;
 	for (;;)
 	{
-		// If player is not alive, or model is not set, wait
-		while (!isAlive(self) || !isDefined(self.pers["savedmodel"]))
-			wait level.frame;
 		wait level.frame;
 
-		// Link tag to head
-		/#
-		println(self.name + " " + self.model); // crash on xmodel/playerbody_german_africa01 | xmodel/playerbody_american_normandy09
-		#/
+		// If player is not alive, or model is not set, wait
+		if (self.sessionstate != "playing" || !isDefined(self.pers["savedmodel"]))
+		{
+			dead = true;
+			continue;
+		}
+
+		// We can call linkto() after 1 frame or it will cause runtime error "failed to link entity since parent model 'xmodel/playerbody_german_africa01' is invalid"
+		if (dead)
+		{
+			dead = false;
+			continue;
+		}
+
+		// Link tags
 		self.headTag unlink();
 		self.headTag linkto (self, "J_Head",(0,0,0),(0,0,0));
 
-		// ^ Crash error:  failed to link entity since parent model 'xmodel/playerbody_american_normandy09' is invalid
-
-		// Wait untill is alive
-		while (isAlive(self))
-			wait level.fps_multiplier * 1;
+		self.pelvisTag unlink();
+		self.pelvisTag linkto (self, "pelvis",(0,0,0),(0,0,0));
 	}
 }
 
@@ -90,6 +111,8 @@ DeleteTags()
 {
 	if (isDefined(self.headTag))
 		self.headTag delete();
+	if (isDefined(self.pelvisTag))
+		self.pelvisTag delete();
 }
 
 
@@ -102,8 +125,14 @@ DeleteTags()
 getEyeOrigin()
 {
 	origin = self getOrigin();
-	head = self.headTag getOrigin();
+
+	if (self.sessionteam == "spectator")
+	{
+		return origin + (0, 0, 8);
+	}
+
 	angles = self getPlayerAngles();
+	head = self.headTag getOrigin();
 
 	// Vector of head
 	diff = head - origin;
@@ -220,8 +249,8 @@ getLookingAtPosition()
 
 GetEntityByClientId(id)
 {
-	if (id == -1)
-		return;
+	if (id <= -1)
+		return undefined;
 
 	players = getentarray("player", "classname");
 	for(i = 0; i < players.size; i++)
@@ -235,3 +264,132 @@ GetEntityByClientId(id)
 	}
 	return undefined;
 }
+
+
+
+/*
+Error produced when invalid name of tag is used via function link():
+
+Models:
+0: 'playerbody_british_africa03'
+79: 'head_british_macgregor'
+107: 'helmet_british_afrca'
+
+Bones:
+Bone 0: 'tag_origin'
+Bone 1: 'j_mainroot'
+Bone 2: 'pelvis'
+Bone 3: 'j_hip_le'
+Bone 4: 'j_hip_ri'
+Bone 5: 'torso_stabilizer'
+Bone 6: 'j_spine1'
+Bone 7: 'back_low'
+Bone 8: 'j_back_equip_le'
+Bone 9: 'j_back_equip_ri'
+Bone 10: 'j_hiptwist_le'
+Bone 11: 'j_hiptwist_ri'
+Bone 12: 'j_knee_le'
+Bone 13: 'j_knee_ri'
+Bone 14: 'j_leftspade'
+Bone 15: 'j_shorts_le'
+Bone 16: 'j_shorts_lift_le'
+Bone 17: 'j_shorts_lift_ri'
+Bone 18: 'j_shorts_ri'
+Bone 19: 'j_spade'
+Bone 20: 'j_spine2'
+Bone 21: 'back_mid'
+Bone 22: 'j_ankle_le'
+Bone 23: 'j_ankle_ri'
+Bone 24: 'j_spine3'
+Bone 25: 'back_up'
+Bone 26: 'j_ball_le'
+Bone 27: 'j_ball_ri'
+Bone 28: 'j_spine4'
+Bone 29: 'j_clavicle_le'
+Bone 30: 'j_clavicle_ri'
+Bone 31: 'j_neck'
+Bone 32: 'neck'
+Bone 33: 'j_head'
+Bone 34: 'head'
+Bone 35: 'j_shoulder_le'
+Bone 36: 'j_shoulder_ri'
+Bone 37: 'j_elbow_bulge_le'
+Bone 38: 'j_elbow_bulge_ri'
+Bone 39: 'j_elbow_le'
+Bone 40: 'j_elbow_ri'
+Bone 41: 'j_shouldertwist_le'
+Bone 42: 'j_shouldertwist_ri'
+Bone 43: 'j_wrist_le'
+Bone 44: 'j_wrist_ri'
+Bone 45: 'j_wristtwist_le'
+Bone 46: 'j_wristtwist_ri'
+Bone 47: 'j_index_le_1'
+Bone 48: 'j_index_ri_1'
+Bone 49: 'j_mid_le_1'
+Bone 50: 'j_mid_ri_1'
+Bone 51: 'j_pinky_le_1'
+Bone 52: 'j_pinky_ri_1'
+Bone 53: 'j_ring_le_1'
+Bone 54: 'j_ring_ri_1'
+Bone 55: 'j_thumb_le_1'
+Bone 56: 'j_thumb_ri_1'
+Bone 57: 'tag_weapon_left'
+Bone 58: 'tag_weapon_right'
+Bone 59: 'j_index_le_2'
+Bone 60: 'j_index_ri_2'
+Bone 61: 'j_mid_le_2'
+Bone 62: 'j_mid_ri_2'
+Bone 63: 'j_pinky_le_2'
+Bone 64: 'j_pinky_ri_2'
+Bone 65: 'j_ring_le_2'
+Bone 66: 'j_ring_ri_2'
+Bone 67: 'j_thumb_le_2'
+Bone 68: 'j_thumb_ri_2'
+Bone 69: 'j_index_le_3'
+Bone 70: 'j_index_ri_3'
+Bone 71: 'j_mid_le_3'
+Bone 72: 'j_mid_ri_3'
+Bone 73: 'j_pinky_le_3'
+Bone 74: 'j_pinky_ri_3'
+Bone 75: 'j_ring_le_3'
+Bone 76: 'j_ring_ri_3'
+Bone 77: 'j_thumb_le_3'
+Bone 78: 'j_thumb_ri_3'
+Bone 79: 'j_spine2'
+Bone 80: 'j_spine3'
+Bone 81: 'j_spine4'
+Bone 82: 'j_neck'
+Bone 83: 'j_head'
+Bone 84: 'j_brow_le'
+Bone 85: 'j_brow_ri'
+Bone 86: 'j_cheek_le'
+Bone 87: 'j_cheek_ri'
+Bone 88: 'j_eye_lid_bot_le'
+Bone 89: 'j_eye_lid_bot_ri'
+Bone 90: 'j_eye_lid_top_le'
+Bone 91: 'j_eye_lid_top_ri'
+Bone 92: 'j_eyeball_le'
+Bone 93: 'j_eyeball_ri'
+Bone 94: 'j_head_end'
+Bone 95: 'j_jaw'
+Bone 96: 'j_levator_le'
+Bone 97: 'j_levator_ri'
+Bone 98: 'j_lip_top_le'
+Bone 99: 'j_lip_top_ri'
+Bone 100: 'j_mouth_le'
+Bone 101: 'j_mouth_ri'
+Bone 102: 'tag_eye'
+Bone 103: 'j_chin_skinroll'
+Bone 104: 'j_helmet'
+Bone 105: 'j_lip_bot_le'
+Bone 106: 'j_lip_bot_ri'
+Bone 107: 'j_helmet'
+
+Part duplicates:
+79 ('j_spine2') -> 20 ('j_spine2')
+80 ('j_spine3') -> 24 ('j_spine3')
+81 ('j_spine4') -> 28 ('j_spine4')
+82 ('j_neck') -> 31 ('j_neck')
+83 ('j_head') -> 33 ('j_head')
+107 ('j_helmet') -> 104 ('j_helmet')
+*/
